Delete comment is working good. And also tacking feeds for it.

+++ src/Controller/TasksController.php
@@ -418,7 +418,9 @@ class TasksController extends AppController
 
         $isTrashed = false;
         if ($this->Attachments->delete($attachment)) {
-            unlink(WWW_ROOT . 'img/attachments/' . $attachment->path);
+            if(file_exists(WWW_ROOT . 'img/attachments/' . $attachment->path)){
+                unlink(WWW_ROOT . 'img/attachments/' . $attachment->path);
+            }
             $isTrashed = true;
         }
